+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### Version last
 - 支持含有枚举的注释生成相关java/kotlin代码
 - 现在你可以在build.gradle.kts中把现有依赖upsert到libs.version.toml ,只需alt+回车 
+- 在实体类上alt+回车基于javadoc注释添加swagger注解
 
 #### Version 1.6
 - 新增结构化输出能力，呼出generate上下文，将自然语言转为当前类对应的Json格式。